Fix for the following:
If you add 2 or more aggregators to a Grafana panel, delete the first, the other will show it as having the interval and alignment from the first aggregator rather than the one you just entered.

The fix is to directly call the getVisibleParameters() method from the controller. I believe this was not working because the UI was not refreshing the controller state ($scope.apply() not being called).

Also, fixed autoValue to correctly set the sampling unit. I probably broke this in my last set of changes.