fix a bug where EnumAggregatorParameter could not be cloned
When adding a new aggregator, the controller calls 'fromObject' to
initialize a fresh instance from a given one. The constructor parameter
'type' is expected to be an Enum name so that the allowed values for
that field can be populated.

Once this is done however, the EnumAggregatorParameter itself saves a
string to its 'type' field. That is, this name is overloaded and was
being misused for two different purposes.